refactor : 테스트코드 수정
